// src/models/Client.ts
export interface ClientType {
  email: string
  tipo_identificacion_id: number
  numero_identificacion_cliente: string
  nombre_cliente: string
  apellido_cliente: string
  password: string
  telefono_cliente: string
  direccion_entrega_cliente: string
  fecha_nacimiento_cliente: string
  ciudad_id: number
}